Architecture Document for Data-Collection Framework
The purpose of this document is to provide insight into our proposed framework so that reader could give comments, suggestions. It explains about the architecture, category of the systems / sub-systems, request pay load and response. As a result of reader’s comment/suggestions, we will get inputs on existing features as well as get to know about the points that we have missed